Make curlRequest thread safe.
I.e. move the curl handle out of gAPI into
curlRequest, where we have a static map
with a handle for each separate thread.
When a curlRequest.perform() is called from
some thread, either the thread's handle
is used or, if it does not exist, a new one
is initialized (via curl_easy_initialize...).